rename HathoraSDK to HathoraSDKModule
